Added event test

+++ models/outcome.js
@@ -1,16 +1,16 @@
-module.exports = function(sequelize, DataTypes) {
+module.exports = function(sequelize, DataTypes){
 	return sequelize.define("Outcome", {
-		id : {
-			type : DataTypes.INTEGER.UNSIGNED,
+		id     : {
+			type          : DataTypes.BIGINT.UNSIGNED,
 			autoIncrement : true,
-			allowNull : false,
-			primaryKey : true
+			allowNull     : false,
+			primaryKey    : true
 		},
 		winner : {
-			type : DataTypes.BOOLEAN,
-			allowNull : true,
-			primaryKey : false,
-			unique : false,
+			type         : DataTypes.BOOLEAN,
+			allowNull    : true,
+			primaryKey   : false,
+			unique       : false,
 			defaultValue : null
 		}
 	});
